Fix missing audio artwork metadata on multiple exports
When the metadata filter is applied on an `Audio` media, the `artwork`
parameter was being unset from the metadata array.

Saving the same `Audio` instance into multiple formats generated the
correct commands for the first export, but the subsequent commands were
missing the `artwork` parameter.

This commit fixes this issue by copying (by value) the `metaArr`
property to a local variable each time the filter is applied.