Output full error from encoding failure
Following up the new available methods on ExecutionFailureException to get the error output, this commit changes the error output when a video or audio encoding fails. To ensure compatibility with older versions of the BinaryDriver, a new exception class was created called CommandExecutionFailure, which receives as argument an instance of ExecutionFailureException, and checks if the instance has the method to get the full error output, in which case it appends the error output to the message. Otherwise it defaults to the generic "Encoding failed" message. Two tests were created to ensure that the full error output is obtained when a video or audio encoding fails.