Fix Stream::getDimensions to accommodate rotated videos
When a video is rotated with a `displaymatrix`, notably when recorded on an iPhone, `ffprobe` returns the dimensions with the rotation applied. So e.g. a portrait video of 720x1280 is reported to have a with of 1280 and height 720.

This gives problems with the `ResizeFilter`, because it will provide a wrong scaling.

This change checks whether there is a `rotation` set in the `side_data_list` and if it is plus or minus 90 or 270, it returns the dimensions _flipped_.

There is probably a more robust way to check the rotation, but I don't know which way...